+++ routes/prodcut.js
@@ -1,6 +1,6 @@
 const { store, all_products_details } = require('../controller/productController');
 const validationError = require('../helper/validationError');
-const { productStoreValidation } = require('../request/productValidation/productStoreValidation');
+const { productStoreValidation } = require('../request/productValidation/productValidation');
 
 const router = [
     {
@@ -11,7 +11,7 @@ const router = [
             handler: all_products_details,
             description: "all products details api",
             notes: 'all products details api',
-            tags: ['auth']
+            tags: ['products']
         }
     }, {
         method: 'POST',
@@ -21,7 +21,21 @@ const router = [
             handler: store,
             description: "store product api",
             notes: 'store product api',
-            tags: ['auth'],
+            tags: ['products'],
+            validate: {
+                payload: productStoreValidation,
+                failAction: validationError
+            }
+        }
+    }, {
+        method: 'PUT',
+        path: '/edit',
+        options: {
+            auth: 'jwt',
+            handler: update,
+            description: "update product api",
+            notes: 'update product api',
+            tags: ['products'],
             validate: {
                 payload: productStoreValidation,
                 failAction: validationError
